pyproject.toml
[project]
name = "airflow-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.15",
    "dotenv>=0.9.9",
    "pandas>=2.3.3",
    "prophet>=1.2.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0",
]
